+++ src/main/java/com/example/holidayswap/service/payment/ILoggingService.java
@@ -3,5 +3,5 @@ package com.example.holidayswap.service.payment;
 import com.example.holidayswap.domain.entity.payment.EnumPaymentStatus;
 
 public interface ILoggingService {
-    public void saveLog(long fromID, long toID, Long amount, EnumPaymentStatus.BankCodeError resultCode, String detail, Double fromBalance, Double toBalance,Double commission);
+    public void saveLog(long fromID, long toID, Double amount, EnumPaymentStatus.BankCodeError resultCode, String detail, Double fromBalance, Double toBalance,Double commission);
 }
